+++ JavaModules/idispatch-gis/idispatch-gis-domain/src/main/java/net/pkhapps/idispatch/gis/domain/model/identity/MunicipalityIdConverter.java
@@ -8,5 +8,5 @@ import javax.persistence.Converter;
  * JPA attribute converter for {@link MunicipalityId}.
  */
 @Converter(autoApply = true)
-public class MunicipalityIdConverter extends DomainObjectIdConverter<Long, MunicipalityId> {
+public class MunicipalityIdConverter extends DomainObjectIdConverter<Integer, MunicipalityId> {
 }
